Changes to SBSGenericDetector::DecodeTDC
Need to make changes to time order the TDC hit data, since
from the VETROC the hits are not in time order.

Added structure TDCHits to SBSGenericDetector.h

In DecodeTDC :
Fill vector of TDCHits with edge and rawtime of hits
Sort TDCHits vector in ascending order of time
Use TDCHits vector when processing the TDC hits.