Modify SBSGenericDetector.cxx and header
In ReadDatabase default to
Set F1 rollover as fF1_RollOver = 64964
and fF1_TimeWindow = 8000

If DecodeTDC is processing a signal from a reference time channel
FOr SBSModeTDC::kTDCSimple
If the model is 6401 (Lo-Res F1) then sets TrigTime = evdata.GetRawData
and sets the LeadingEdge = 0.
For all other TDC models set TrigTime=0 and
sets LeadingEdge = evdata.GetRawData
Calls ProcessSimple with the TrigTime in 4th argument.

If DecodeTDC is processing a signal from the detector,
then DecodeTDC first determines which reference time TDC hit to
use ( if case there are multiple TDC hits in the reference time
for that event). If the TDC model is 6401 (Lo-res F1) then it retrieves the
trigtime with GetTrigTime method. If the trigtime is larger
than the raw reference time then the F1_RollOver is added to the
raw reference time set to  a variable tempdata. This used to determine
which TDC hit time is closest to the expected time.
The raw tdc time ( no trig time subtraction) is
saved as the good reference time as the  variable reftime

If the SBSModeTDC::kTDCSimple and TDC model 6401,
then for the detector TDC hits check if absolute difference
between the Detector time and reference time is larger
than time window which means that there is rollover.
Usually the reference time should be larger than the
detector time, so if the reference time has the rollover added.
But it could be a random detector time which is after the
reference time, then the rollover is added to the reference time.
ProcessSimple is called with the raw reference time subtracted
from the detector time.

In CoarseProcess method, for the reference time when
in kTDCSimple mode , then the leading edge value is calculated
subtracting off the trigger time from the raw TDC time.