Major cleanup of unused code and TTree outputs from SBSScalerHelicity.
There is no change to the behavior of the "scalhel.hel" variable in the
physics "T" tree.  That remains the best way to tag the physics events
with the helicity state.

When we wrote SBSScalerHelicity originally, it was derived from previous
scaler helicity analysis classes, and we had not bothered to entirely
cleanup functions, data members, and "TShel" TTree output veriables
which had become superceeded.
That made it kind of difficult to remember how to actually plot the (raw)
charge asymmetry or otherwise manipulate the "TShel" TTree.

I have now removed unused functions and data members from the class, and
substantially cleaned up the branches in the "TShel" TTree.
To plot the raw asymmetry for the second scaler channel you would do:
    TShel->Draw("Lhel.Diff.Ch1/Lhel.Yield.Ch1",
                "Lhel.ring.PattPhase==3 && Lhel.ring.seedActual!=0")
The cuts on Lhel.ring.PattPhase and Lhel.ring.seedActual are to ensure
that we are looking at the last event of each pattern, and only
considering patterns after the "true helicty" seed has been correctly
calculated.

NOTE:  Channel names are not included yet, so you have to know what
channel number corresponds to the BCM you want to plot.

NOTE:  The pedestal correction is not yet done in the code.  Because
the scaler results for each pattern are presented as the HC difference
and the yield, it is possible to apply the pedestal corrections to the
yield within a ROOT script.
This would look something like the following, where I have take the
pedestal to be 200.0 for this example:
    TShel->Draw("Lhel.Diff.Ch1/( Lhel.Yield.Ch1 - 200. )",
                "Lhel.ring.PattPhase==3 && Lhel.ring.seedActual!=0")